Fix change responsive stage from text options
