+++ src/editor/responsive-selector/index.js
@@ -10,6 +10,7 @@ import { createBlock } from '@wordpress/blocks';
  */
 import Button from '../../components/button';
 import Icon from '../../components/icon';
+import MaxiStyleCardsEditorPopUp from '../style-cards';
 
 /**
  * External dependencies
@@ -30,8 +31,6 @@ import {
 	cloudLib,
 } from '../../icons';
 
-import MaxiStyleCardsEditorPopUp from '../style-cards';
-
 /**
  * Components
  */
@@ -41,8 +40,9 @@ const ResponsiveSelector = props => {
 	const { insertBlock } = useDispatch('core/block-editor');
 
 	const { deviceType, breakpoints } = useSelect(select => {
-		const { receiveMaxiDeviceType, receiveMaxiBreakpoints } =
-			select('maxiBlocks');
+		const { receiveMaxiDeviceType, receiveMaxiBreakpoints } = select(
+			'maxiBlocks'
+		);
 
 		return {
 			deviceType: receiveMaxiDeviceType(),
@@ -59,43 +59,13 @@ const ResponsiveSelector = props => {
 	const classes = classnames('maxi-responsive-selector', className);
 
 	const setScreenSize = size => {
-		const editorWrapper = document.querySelector('.editor-styles-wrapper');
-		const winHeight = window.outerWidth;
-		const responsiveWidth =
-			(size === 'general' && 'none') ||
-			(size === 'xxl' && 2000) ||
-			breakpoints[size];
-
-		editorWrapper.setAttribute(
-			'maxi-blocks-responsive',
-			size !== 'general' ? size : ''
-		);
-		editorWrapper.setAttribute(
-			'maxi-blocks-responsive-width',
-			responsiveWidth
-		);
-
-		if (size === 'general') {
-			editorWrapper.style.width = '';
-			editorWrapper.style.margin = '';
-
-			setMaxiDeviceType('general');
-		} else {
-			const xxlSize = 2000; // Temporary value, needs to be fixed
-
+		const xxlSize = 2000; // Temporary value, needs to be fixed
+		if (size === 'general') setMaxiDeviceType('general');
+		else
 			setMaxiDeviceType(
 				size,
 				size !== 'xxl' ? breakpoints[size] : xxlSize
 			);
-
-			if (size !== 'xxl')
-				editorWrapper.style.width = `${breakpoints[size]}px`;
-			else editorWrapper.style.width = `${xxlSize}px`;
-
-			if (winHeight > breakpoints[size])
-				editorWrapper.style.margin = '36px auto';
-			else editorWrapper.style.margin = '';
-		}
 	};
 
 	return (
