+++ src/extensions/store/reducer.js
@@ -1,5 +1,35 @@
 import controls from './controls';
 
+const breakpointResizer = (size, breakpoints) => {
+	const editorWrapper = document.querySelector('.editor-styles-wrapper');
+	const winHeight = window.outerWidth;
+	const responsiveWidth =
+		(size === 'general' && 'none') ||
+		(size === 'xxl' && 2000) ||
+		breakpoints[size];
+
+	editorWrapper.setAttribute(
+		'maxi-blocks-responsive',
+		size !== 'general' ? size : ''
+	);
+	editorWrapper.setAttribute('maxi-blocks-responsive-width', responsiveWidth);
+
+	if (size === 'general') {
+		editorWrapper.style.width = '';
+		editorWrapper.style.margin = '';
+	} else {
+		const xxlSize = 2000; // Temporary value, needs to be fixed
+
+		if (size !== 'xxl')
+			editorWrapper.style.width = `${breakpoints[size]}px`;
+		else editorWrapper.style.width = `${xxlSize}px`;
+
+		if (winHeight > breakpoints[size])
+			editorWrapper.style.margin = '36px auto';
+		else editorWrapper.style.margin = '';
+	}
+};
+
 const reducer = (
 	state = {
 		settings: {},
@@ -41,6 +71,7 @@ const reducer = (
 				deviceType: action.deviceType,
 			};
 		case 'SET_DEVICE_TYPE':
+			breakpointResizer(action.deviceType, state.breakpoints);
 			return {
 				...state,
 				deviceType: action.deviceType,
